Model metadata functionality
Added functions to allow us to store metadata about model runs, so we do not have to memorise everything ourselves. Model parameters will be read out of the source code files, and parameters are extracted using regex. Parameters to be tracked can be listed in the params.json file - more will be added later.